// src/app/openApis/AssuranceServicesManagementAPI/models/action-specification-update.ts
/* tslint:disable */
import { ActionParam } from './action-param';

/**
 * An ActionSpecification is an entity that describes an action to perform on certain entities.
 */
export interface ActionSpecificationUpdate {

  /**
   * Description of the entity
   */
  description?: string;

  /**
   * The name of the entity
   */
  name?: string;
  params?: Array<ActionParam>;
}